split files and removed experimental code
