+++ config.py
@@ -1,21 +1,22 @@
 UNIT = 40
-LEFT_BORDER = 0
-RIGHT_BORDER = 10
-UPPER_BORDER = 22
-LOWER_BORDER = 0
 GRAVITY_INTERVAL = 1
 
-O_PIECE = [(4, 20), (5, 20), (4, 21), (5, 21)]
-O_PIECE_ROTATION = (5, 21)
-I_PIECE = [(3, 20), (4, 20), (5, 20), (6, 20)]
-I_PIECE_ROTATION = (5, 20)
-J_PIECE = [(3, 21), (3, 20), (4, 20), (5, 20)]
-J_PIECE_ROTATION = (4.5, 20.5)
-L_PIECE = [(3, 20), (4, 20), (5, 20), (5, 21)]
-L_PIECE_ROTATION = (4.5, 20.5)
-S_PIECE = [(3, 20), (4, 20), (4, 21), (5, 21)]
-S_PIECE_ROTATION = (4.5, 20.5)
-Z_PIECE = [(3, 21), (4, 21), (4, 20), (5, 20)]
-Z_PIECE_ROTATION = (4.5, 20.5)
-T_PIECE = [(3, 20), (4, 20), (5, 20), (4, 21)]
-T_PIECE_ROTATION = (4.5, 20.5)
+PIECES = {
+    'O_PIECE': [(4, 20), (5, 20), (4, 21), (5, 21)],
+    'I_PIECE': [(3, 20), (4, 20), (5, 20), (6, 20)],
+    'J_PIECE': [(3, 21), (3, 20), (4, 20), (5, 20)],
+    'L_PIECE': [(3, 20), (4, 20), (5, 20), (5, 21)],
+    'S_PIECE': [(3, 20), (4, 20), (4, 21), (5, 21)],
+    'Z_PIECE': [(3, 21), (4, 21), (4, 20), (5, 20)],
+    'T_PIECE': [(3, 20), (4, 20), (5, 20), (4, 21)]
+}
+
+PIECE_ORIGINS = {
+    'O_PIECE': (5, 21),
+    'I_PIECE': (5, 20),
+    'J_PIECE': (4.5, 20.5),
+    'L_PIECE': (4.5, 20.5),
+    'S_PIECE': (4.5, 20.5),
+    'Z_PIECE': (4.5, 20.5),
+    'T_PIECE': (4.5, 20.5)
+}
