+++ python-tetris.py
@@ -1,40 +1,35 @@
 import pyglet
 import config
-import piece
+import map
 
 
 class Window(pyglet.window.Window):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.piece = piece.Piece(config.T_PIECE,
-                                 config.T_PIECE_ROTATION)
+        self.map = map.Map(int(self.width/config.UNIT),
+                           int(self.height/config.UNIT))
 
     def on_draw(self):
         self.clear()
-        self.piece.render()
+        self.map.render()
 
     def on_text_motion(self, motion):
-        if motion == pyglet.window.key.MOTION_LEFT:
-            self.piece.move_left()
-        if motion == pyglet.window.key.MOTION_RIGHT:
-            self.piece.move_right()
-        if motion == pyglet.window.key.MOTION_DOWN:
-            self.piece.move_down()
+        self.map.move(motion)
 
     def on_key_press(self, symbol, modifier):
         if symbol == pyglet.window.key.SPACE:
-            self.piece.hard_drop()
-        if symbol == pyglet.window.key.MOTION_UP:
-            self.piece.rotate_cw()
-        if symbol == pyglet.window.key.Z:
-            self.piece.rotate_ccw()
+            self.map.hard_drop()
+        elif symbol == pyglet.window.key.MOTION_UP:
+            self.map.rotate_cw()
+        elif symbol == pyglet.window.key.Z:
+            self.map.rotate_ccw()
 
     def piece_gravity(self, dt):
-        self.piece.move_down()
+        self.map.gravity()
 
 
 if __name__ == '__main__':
-    window = Window(10*config.UNIT, 22*config.UNIT,
+    window = Window(400, 880,
                     "Python Tetris", resizable=True)
     pyglet.clock.schedule_interval(window.piece_gravity,
                                    config.GRAVITY_INTERVAL)
